Support for processes module under Solaris
The processes module under Solaris is still in very early development. The files in /proc have been identifier but the code crashed on Solaris SPARC consistently. This commit is purely developmental and meant to be used in order to facilitate the code transfer from one platform to another. Once the module is satisfactory, it will be merged with the main branch.