Processes for Solaris-only module with read-write support
Added write support for Solaris as well as fork rate count based on kstat nthreads count.
This is still a Solaris only version of the processes module. However this version is stable and sends writes a great deal of information about process states. The only thing not yet implemented are the data and code section monitoring.

Also, I have removed all the debugging printf statements.