Added SendPercent configuration variable for the cpu module
Because the values send by Linux machines are in jiffies and the values
send my Solaris machines are in cycles, the two cannot be aggregated correctly
by reporting tools. It makes more sense if all servers, regardless of their OS
send a consistent value. The configuration variable SendPercent forces Solaris
machines to send the cpu load as e number between 0 and 100.